refactor: move RespanSpanAttributes enum to constants per BE convention
BE convention: "All Enums ARE constants, always place in constants.py".
Move enum to span_attributes.py (constants file), re-export from
span_types.py for backend backward compat. Dependency direction is
now correct: types imports from constants, not the reverse.

Also clarify GenAI constants comment — not yet in semconv_ai, defined
locally until upstream includes them.
